xfeatures2d: SURF descriptor: leave the class_id field of KeyPoint unassigned
Presently, all three implementations (CPU, OpenCL, CUDA) of
keypoint detector fill the class_id field of KeyPoint structure
with the value of laplacian. However, this value is neither read
back when computing descriptor, nor used by SURF or any other
descriptor extractor implementations.

Plus, the OpenCV docs suggest that the class_id should be used by
user, for example to group the keypoints. While filling them with
values in keypoint detector and not using them later does not
violate that, it is still unnecessary and confusing.